refactor: remove not used code and clean code
