--- src/shared/constants/index.ts
@@ -1,5 +1,0 @@
-export const PACKAGES = {
- UI: 'ui',
- API: 'api',
- WEB3: 'web3',
-};
